@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace Laminas\ApiTools\Doctrine\Server\Event\Listener;
 
-use Doctrine\Common\Persistence\Mapping\ClassMetadata;
-use Doctrine\Common\Persistence\ObjectManager;
 use Doctrine\Instantiator\InstantiatorInterface;
+use Doctrine\Laminas\Hydrator\DoctrineObject;
 use Doctrine\ORM\Internal\Hydration\AbstractHydrator;
-use DoctrineModule\Stdlib\Hydrator\DoctrineObject;
+use Doctrine\Persistence\Mapping\ClassMetadata;
+use Doctrine\Persistence\ObjectManager;
 use Laminas\ApiTools\Doctrine\Server\Event\DoctrineResourceEvent;
 use Laminas\ApiTools\Doctrine\Server\Exception\InvalidArgumentException;
 use Laminas\EventManager\EventManagerInterface;
