@@ -4,39 +4,18 @@ declare(strict_types=1);
 
 namespace Laminas\ApiTools\Doctrine\Server\Query\CreateFilter;
 
-use Doctrine\Common\Persistence\ObjectManager;
 use DoctrineModule\Persistence\ObjectManagerAwareInterface;
+use DoctrineModule\Persistence\ProvidesObjectManager;
 use Laminas\ApiTools\Rest\ResourceEvent;
 
 abstract class AbstractCreateFilter implements ObjectManagerAwareInterface, QueryCreateFilterInterface
 {
+    use ProvidesObjectManager;
+
     /**
      * @param string $entityClass
      * @param array $data
      * @return array
      */
     abstract public function filter(ResourceEvent $event, $entityClass, $data);
-
-    /** @var ObjectManager */
-    protected $objectManager;
-
-    /**
-     * Set the object manager
-     *
-     * @return void
-     */
-    public function setObjectManager(ObjectManager $objectManager)
-    {
-        $this->objectManager = $objectManager;
-    }
-
-    /**
-     * Get the object manager
-     *
-     * @return ObjectManager
-     */
-    public function getObjectManager()
-    {
-        return $this->objectManager;
-    }
 }
