@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace LaminasTestApiToolsDbMongo\Repository;
 
-use Doctrine\ODM\MongoDB\DocumentRepository;
+use Doctrine\ODM\MongoDB\Repository\DocumentRepository;
 
 class MetaRepository extends DocumentRepository
 {
